backend/requirements.txt
# Core ML
torch==2.5.1
torchvision==0.20.1
transformers==4.47.0
accelerate==1.1.0

# FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
python-multipart==0.0.12

# Image/Video
opencv-python==4.10.0.84
pillow==10.4.0
numpy==1.26.4

# Optional
pydantic==2.9.0

num2words==0.5.14